+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 This is intended as an overview of the major changes
 
-v3.2
+v4.0
 ===
+- Switching to require Java 8
 - Parser instance is now serializable.
 - Added example on using with Apache Flink
 - Added example on using with Apache Beam